Funcion para registrar los iconos de los organismos separado de la ruta y agregado como una funcion en el controlador de organismos
