+++ backend/.gitignore
@@ -98,6 +98,3 @@ Thumbs.db
 
 # Test files
 test-results/
-
-# File img
-src/img/
