+++ backend/src/routes/organism.ts
@@ -63,31 +63,6 @@ router.get(
     }
 });
 
-// Ruta para obtener la imagen del organismo
-router.get(
-  "/icono/:filename",
-  async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const filename = req.params.filename;
-
-      // Construir la ruta absoluta al archivo de imagen
-      const imagePath = path.join(__dirname, "../static/organism", filename);
-
-      // Verificar que el archivo exista
-      if (!fs.existsSync(imagePath)) {
-        res.status(404).json({ message: "Imagen no encontrada" });
-        return;
-      }
-
-      // Enviar el archivo de imagen
-      res.sendFile(imagePath);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
-
-
 router.post(
     "/",
     requirePermission("create_organism"),
@@ -104,23 +79,10 @@ router.post(
                 res.status(400).json({ message: "Icono de organismo requerido." });
                 return;
             }
+            const file = req.file;
 
-            // Ruta de destino para la imagen
-            const ext = path.extname(req.file.originalname);
-            const newFileName = `${name}${ext}`;
-            const destDir = path.join(__dirname, "../static/organism");
-            const destPath = path.join(destDir, newFileName);
+            const icono = await OrganismController.validateImage(name, file);
 
-            //Asegurar que la carpeta exista
-            if (!fs.existsSync(destDir)) {
-                fs.mkdirSync(destDir, { recursive: true });
-            }
-
-            //renombar y mover la imagen de la ruta temporal a la ruta destino
-            fs.renameSync(req.file.path, destPath);
-
-            //guardar la ruta local con el archivo "nuevo"
-            const icono = `api/organisms/icono/${newFileName}`;
             //const icono = `src/img/organism/${newFileName}`;
             const newOrganism = await OrganismController.createOrganism(name, icono);
             res.status(201).json(newOrganism);
